Add online status badge for other contacts
