+++ src/renderer/components/ColorBadge.css
@@ -1,0 +1,6 @@
+.ColorBadge {
+  height: 100%;
+  width: 100%;
+  border-radius: 50%;
+  border: 1px solid white;
+}
