--- src/renderer/components/content-types/contact/ConnectionStatusBadge.css
@@ -1,18 +1,0 @@
-.ConnectionStatus {
-  height: 10px;
-  width: 10px;
-  border-radius: 50%;
-  border: 1px solid white;
-}
-
-.ConnectionStatus--noDevices {
-  background-color: grey
-}
-
-.ConnectionStatus--notConnected {
-  background-color: orange
-}
-
-.ConnectionStatus--connected {
-  background-color: green
-}
