+++ src/renderer/components/content-types/contact/ContactInVarious.css
@@ -31,8 +31,10 @@
   position: relative;
 }
 
-.Contact-connectionStatus {
+.Contact-status {
   position: absolute;
   right: 0;
   bottom: 0;
+  height: 8px;
+  width: 8px;
 }
